Added javadoc for this Ant task.
Annoyance with quoting XML in Javadoc: you have to escape everything.
Fortunately, this was remedied by the {@code ...} tag.

Annoyance with quoting an Ant task in Javadoc: you probably have a path pattern
in there that looks like **/blah. Well, guess what? */ closes the damned
Javadoc comment. Normally you'd work around this by replacing the / with &#47;,
which sucks because it makes the example hard to read inside the source, but is
tolerable. But you can't use that inside a {@code ...} tag. Dooh!

So I have to do this bastard combination of escapes. Yay!