Fixed beans/frank in bindHovered. Added bindStateStyle().
In theory, one would like to Value.map a boolean value to a string which would
evaluate to the on style or the off style depending on the state of the
underlying boolean, and then just have Bindings.bindStyle(Value<String>,
Widget...). However, we don't report the previous value of a value when it
changes, so we would not know which old style to remove when the Value<String>
changed. Other FRP systems do provide the previous value when a value changes,
so perhaps I should take the plunge and add this. I worry that it will mean
updating a zillion Value.Listener implementations, but maybe not. Maybe most of
the FRP stuff we do is to just use Bindings directly which hides the listening
under the hood. That would be nice.