In GWT 2.1.0, KeyPressEvent.getCharCode no longer returns key codes. We should have been using KeyDownEvent.getNativeKeyCode, and now we have to because otherwise shit don't work.
This unfortunately will break all users of EnterClickAdapter because it needs
to be addKeyDownHandlered rather than addKeyPressHandlered. I added a static
bind() method that we can use in lieu of manual adapter addage to avoid
breakage like this in the future. However, we probably won't have to change
things again...