Extracted not-null checking into Preconditions. I had a Preconditions sitting around that I wanted to commit a while ago, but opted not to out of desire not to step on Guava's toes. However, the value of zero dependencies outweighs my desire to avoid duplicating functionality.
In GWT land, incorporating additional dependencies is a pain in the ass.
Incorporating Guava into your GWT app is, in fact, a pain in the ass of
proportions so monumental that the duplicated code would need to get into the
hundreds or thousands of lines before I'd consider it worth burdening users of
gwt-utils with that integration.