Fix javadocs.
This includes a change to show protected members in javadoc. This seems like the right setting for
three rings libraries since we provide a lot of classes that are designed to be extended. But for
some reason the de facto default (probably copied from narya or something), is set to public. So